+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Updated all projects to .NET 6
 * Viewer: Updated Avalonia to 0.10.12
+* Viewer: Removed dependency on deprecated WebRequest APIs
 
 ## v0.11.6 ##